Move Go profiling helpers out of samples/
These are used in cmd as well, so it seemed weird to have
them over in samples/util
