+++ go/util/profile/profile.go
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, version 2.0:
 // http://www.apache.org/licenses/LICENSE-2.0
 
-package util
+package profile
 
 import (
 	"flag"
